Corrected the spacing parameters to accept values in mm, not in pixels.
The --help message stated that the padding, vspace and hspace parameters
took values in mm - and that was probably the intended behavior - but in
reality, those values were applied as pixels. The Change converts them from
mm to pixels using the supplied  resolution.

(Also, some minor code style tweaks and comments that I couldn't be
bothered to put into a separete changeset ;)